Remove outdated test case
This test case definition is no longer really sensible. Decisions about
which log message is shown or not should be done on the handler level.
The logger itself will forward the log message regardless. If the
logging is disabled, that means that all handlers are removed (except
for the `NullHandler`).