Remove type hint from `Camera.get_pixel_format`
The type hint resulted in a mypy error regarding a missing return
statement. In my oppinion this is a false positive since the `for` loop
covers all available pixel formats and therefore a return will always be
encountered. But as this branch is mainly concerned with tidying up the
docstrings this is not a good place to work on that.