Fix `_Tracer.is_log_enabled` logic
The previous implementation was based on an older `Log` implementation.
With the migration to using a more "standard" logging instance for
VmbPy, the logic has to be adjusted. The logger is now considered
enabled, if there are any handler registered that are not `NullHanlder`
instances. The `NullHandler` is registered for a disabled logger to make
sure that our library does not forward any messages to the root logger
to not spam users logs from our API. The user can register their desired
handler or use our convenience function `Log.enable` if they want to
receive log messages. (see e.g. the `create_trace_log.py` example)