Ensure that context decorator always comes first
In the test suite a few test cases validate that the
`RaiseIfOutsideContext` and `RaiseIfInsideContext` decorators generate
`RuntimeError`s if the appropriate context state is not reached. In the
old order where in some places the `TranceEnable` decorator was the
top-most one, is was possible to generated `TypeError`s in those test
cases if logging was enabled because the function signature is
evalueated before the context state is checked. Reordering the
decorators prevents this by making sure that the context state is always
evaluated before a function signature is checked. The overall behaviour
of the decorated functions should not be impacted by this.