Stabilize test `tearDown` for some cases
In some test cases the camera is closed during the test case. During
camera close the cameras internal `LocalDevice` is set to `None`. If the
cameras close method is called again, this leads to an error because the
method tries to call a member function on this `None` that does not
exist. Since the user is not supposed to handle calls to camera open and
close it was decided that the behaviour is save since under correct use
of the api close will not be called multiple times. The test case
however *did* call close multiple times. This `try... except` construct
is not ideal, but it prevents the test suite from hangig and is deemed
acceptable.