Remove `:` from `PixelFormat` comments
The `:` lead sphinx to believe that everything befor the `:` should be
considered a "Type" and only the text after `:` was a general docstring
regarding the enum entry. I assume this is because sphinx tries to parse
the line with python type annotations in mind.

Removing the `:` from the comments does not reduce readability (in fact
it improves consistencty) and leads to the expected documentation
layout.