Add proof-of-concept for user supplied buffers
By allowing the user to supply a buffer into which the image data should
be written, we can completely avoid memory allocations in the
`convert_pixel_fomrat` method. The user is responsible for providing an
appropriate buffer by e.g. generating a numpy array with correct size
and passing the `array.data` member as `destination_buffer` parameter.
The size and layout of the data buffer must be appropriate to write the
resulting image data into it.