Add `frame` parameter to `wait_for_frames`
and rename method to `wait_for_frame` as it will now no longer try to
call `VmbCaptureFrameWait` for all frames in context but only the passed
one. This makes it more consistent in signature to the `queue_frame`
method that is used in conjunction with it. It might make sense to add
some test to these methods to ensure that only frames from the current
context are passed to it.