skip tests that fail for SimTL
These tests are expected to fail for SimTL but not for real cameras:
In SimTL, Acquisition Mode only has "Continuous" which is OK, but generally "SingleFrame" and "MultiFrame" are expected too.
"DeviceScanType" has Areascan and Linescan, but in real cameras only "Areascan" exists.
A frame is expected to arrive more than 1 second after acquisition is started only for real cameras.