Reduce output of `chunk_callback`
This reduces the work done in the chunk callback function and makes the
output easier to follow during streaming. Also the output is now more
consistent with the output from the VmbC chunk example.

In some cases it was possible, that the old `chunk_callback`
implementation took too long and blocked the frame callback function.
This prevented the frame buffer from being returned to VmbC for future
frames in time to have a stable stream. This means that a lot of
incomplete frames were received.