Fix unittest skip decorator
The unittest skip decorator condition is evaluated when the class is
generated. With the old implementation at that point the `test_cam_id`
was always an empty string meaning that the condition always evaluated
to `False`. By now using the environment variable as initial guess the
skipping should work fine as long as the `test_cam_id` is supplied as
environment variable. If the user supplies the id as a command line
parameter with `run_tests.py` this will likely not work reliably.