Fix conversion if user supplies buffer and formats are equal
If the user supplies a buffer to write the image data to and the target
format is the same as the input format the copy of the data buffer is
now performed manually. The call to `deepcopy` still sets
`result._buffer` in `Frame.__deepcopy__`, but it will now take the
object that is provided by the existing entry in the `memo` dict we are
passing in. This basically tells the `deepcopy` function "this field is
already copied and you can just update the reference if you need it
again".