Roll state machine further back to guarantee full requeue
This change moves the state machine another state further back between
frame acquisitions. This means that after a frame has been received
`VmbCaptureFrameQueueFlush` will be called. Upon starting the next loop
iteration this will requeue all frames in the state machines context
which is necessary for systems where the `StreamAnnounceBufferMinimum`
feature reports a value >1. Before this change it was not possible to
record multiple frames synchronously because only on the first frame
were enough frames queued. On the second iteration of the loop only a
single frame would be queued which prevented stable frame transmission.