Add initial version of new example
This is meant to showcase how `convert_pixel_format` might be used. The
example also tries to give a proper explanation how a user might
allocate an appropriate object that can be used as `destination_buffer`
and how to use it correctly. The example can also be helpful in
showcasing the reduced garbage collector work if the
`destination_buffer` is used.