Make tests run better with real cameras
Most of the existing tests did not really need `Rgb8` as pixel format.
It is mostly needed for the test that actually verifies pixel values
because channel order is easy to reverse in numpy. For other tests
`Mono8` is fine and is available on most (all?) devices. If `Rgb8` is
not available for the device used to execute the tests the corresponding
test case is skipped and an appropriate reason is given.