Fix conversion for `target_fmt == fmt` case
This removes the early return from the function and instead of
performing a simple deepcopy of the frame it actually constructs a new
frame instance and performs a copy of the data in step 6 of the method.

This is slightly more work than the old implementation, but it works
correctly when the user supplies a destination buffer that has the
expected size for the image data, but is not large enough to also hold
the possibly existing chunk data.