Remove `[...].pop()` pattern to prevent `IndexError`
The error was never encountered in actual applications, but the code
pattern is still not ideal. This prevents the exception in the first
place and even handles an error case where a `Reachable` or
`Unreachable` event might be triggered for devices that were not yet
known to `VmbSystem`. In such a case the device is now added to the list
of known devices.