Fix `imageData` pointer handling in frame conversion
After conversion the image data will always start at the beginning of
the buffer sinceany chunk data that might have existed there before is
not part of the conversion result.

Similarly the data pointer handed to the image transform library should
always point to the actual start of the image data, not the start of the
buffer.