Add test for enum string representation
With Python 3.11 the string representation of `enum.IntEnum` changed.
This means that for our enums using that type the enums value (i.e. the
integer value) is printed by default instead of the entries name. The
desired behaviour here is however to print the name for better
readability. The `IntEnum` type was chosen because it has some
guarantees that the underlying value was always an integer which made
the ctypes passing a little safer.