Make `FrameConsumer` run in main thread
On Mac OS there is a known issue, that means that OpenCV GUI functions
may only be called from the applications main thread. With the old
implementation the `FrameConsumer`, which did all the display calls, was
in a separate thread. By not inheriting from `Treading.Thread` for the
`FrameConsumer` and `MainThread` classes, we can actually execute the
consumer in the main thread and hopefully fix the issue.