Fix flake8 complaint about bare except
A bare `except` is identical to `except BaseException`. This can be
problematic because it would also catch things like `KeboardInterrupt`
which is an exception in Python. For situations where all errors should
be caught `except Exception` is recommended instead.

For more details see https://www.flake8rules.com/rules/E722