Implement reset and load of default user set
The default user set will be loaded if the test suite is executed via
the provided `run_tests.py` file or by calling `Tests/runner.py`
manually. The default user set will *not* be loaded when auto test
discovery approaches are used such as `python -m unittest ...`. For
auto test discovery there is no mechanism that resembles the global
fixture we would need to implement this.