Add test for frame status to conversion tests
If the frame is not complete, it might mean that no data was written to
it and that all pixels still contain `0x00` as value from the buffer
initialization. This makes the assertions further down in the test
cases useless where it is asserted, that data was written to the
expected buffers.