Skip unsupported Image Transform call for some cases
The function `VmbSetImageInfoFromInputImage` only supports 8 and 16 bit
layouts. If the user has a packed image and wants to call
`convert_pixel_format` on it with the *same* packed format, the
conversion should not take place and instead a simple copy should be
performed. In that case it does not matter, that the Image Transform
library does not support the conversion *to* a packed format, as it
will not have to perform that conversion anyway.