Add timeout parameter to `get_frame` calls
The test class sets the exposure time to a very high value (depending
on camera model). This might lead to timeout errors if the exposure time
is longer than the default frame timeout. This commit sets the
timeout value larger than the exposure time to prevent this.

Additionally this commit also resets the long exposure time the class
sets to a smaller value so that subsequent streaming tests do not take
an eternity due to the long exposure.