Fix resolution of forward declared objects
In Python 3.14 the resolution of forward declared types has changed. It
seems like the current implementation of our runtime type checking is
not quite the way that 3.14 expects this to work because even though
the check is done at runtime when all classes should be available, the
call to `get_type_hints` raises a `NameError` when e.g.
`Stream.start_streaming` is called ("`Camera` is not defined"). Adding
the `vmbpy` module as localns to the call fixes these issues but it
might make sense to revisit the type checking logic at some time to
determine why exactly the resolution does not work.

References:
- https://docs.python.org/3.14/whatsnew/3.14.html#whatsnew314-deferred-annotations
- https://docs.python.org/3.14/whatsnew/3.14.html#whatsnew314-porting-annotations