Fix possible VmbCError in camera discovery
If a device becomes available in the short time between the two calls
to VmbCamerasList, the second call may return error code
VmbErrorMoreData. In our python-c interface this results in a VmbCError
being raised as exception. This change will catch that exception and
silently retry the calls until the second call no longer returns that
error code.