Ensure that VmbShutdown is always called
Without the `try...catch` it is possible to get an exception in e.g. the
camera discovery part of the `_startup()` method. This exception would
lead mean that the context manager is not properly entered and therefore
the `__exit__` method is never called. Having the error handling in
`__enter__` ensure that even in error cases we do not end up with a
call to `VmbStartup` that is not paired to a call to `VmbShutdown`.